+++ configurations/scripts/data/careerApplicationData.js
