+++ configurations/scripts/data/newsLetterSubscriptionData.js
